+++ CHANGELOG.md
@@ -4,7 +4,16 @@ Inspired by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Note: this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.0.0] - Unreleased
+## [4.0.0] - 2026-05-05
+### Added
+- Generalized `belongs_to` foreign keys to always match the primary key they point at, including
+  in HABTM intersection tables. (Previously this was just a special case that matched up
+  :integer vs :bigint.)
+
+### Removed
+- Drop support for Rails 6.x. Minimum supported Rails is now 7.0.
+
+## [3.0.0] - 2025-04-08
 ### Changed
 - The `timestamps` DSL method to create `created_at` and `updated_at` columns now defaults to `null: false` for `datetime` columns
 - The `timestamps` DSL method to allow additional options to be passed to the `datetime` fields